refactor: replace context files into context folder
